+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+### 20.0.0
+
+- upgrade deps
+- compat composer 2.0
+
+
+### 2.0.0
+
+- compat nc20
+
+
+### 1.5.2
+
+- compat nc19
+- compat ES 7.7+
+
+
+### 1.5.1
+
+- more log during search request
+
 
 ### 1.5.0 (nc18)